style:#209 ai 추천 버튼 디자인 구현 및 ai 추천 버튼 컴포넌트 분리
